Add basic plotting functions for different vectorisation  methods
