+++ gtda/plotting/__init__.py
@@ -3,12 +3,18 @@ transformers' output."""
 
 from .point_clouds import plot_point_cloud
 from .homology import plot_diagram
-from .diagrams import plot_betti_curves, plot_betti_surfaces, plot_landscapes
+from .diagrams import plot_betti_curves, plot_betti_surfaces, plot_landscapes,\
+    plot_persistence_image, plot_silhouettes, plot_heat_kernel
+from .image import ImagePlotterMixin
 
 __all__ = [
     'plot_point_cloud',
     'plot_diagram',
     'plot_betti_curves',
     'plot_betti_surfaces',
-    'plot_landscapes'
+    'plot_landscapes',
+    'plot_persistence_image',
+    'plot_silhouettes',
+    'plot_heat_kernel',
+    'ImagePlotterMixin',
 ]
