+++ gtda/plotting/_plot.py
