+++ gtda/plotting/image.py
